chore: Rename hook file
