+++ mobile/src/hooks/useLayoutConfigs.ts
